refactor: extract bp_alerts + bp_budget to routes/alerts.py
Phase 5.6 of the incremental modularisation. Moves the 6 endpoints on
bp_budget and the 10 endpoints on bp_alerts into a single new module
routes/alerts.py.

bp_budget:
  GET/POST /api/budget/config
  GET      /api/budget/status
  POST     /api/budget/auto-pause
  POST     /api/budget/pause
  POST     /api/budget/resume
  POST     /api/budget/test-telegram

bp_alerts:
  GET/POST /api/alerts/rules
  PUT/DEL  /api/alerts/rules/<rule_id>
  GET      /api/alerts/history
  POST     /api/alerts/history/<int>/ack
  GET      /api/alerts/active
  GET/POST /api/alerts/webhook
  POST     /api/alerts/webhook/test
  GET      /api/alerts/velocity
  GET/POST /api/alert-channels
  POST     /api/alert-channels/test

Pure mechanical move — zero behaviour change. Module-level helpers
(_get_budget_config/_set_budget_config, _get_budget_status,
_pause_gateway/_resume_gateway, _get_alert_rules, _get_alert_history,
_get_active_alerts, _load_alerts_webhook_config/_save_alerts_webhook_config,
_send_webhook_alert/_send_slack_alert/_send_discord_alert,
_compute_velocity_status, _fleet_db/_fleet_db_lock) stay in dashboard.py
and are reached via late "import dashboard as _d".

api_budget_pause now writes to _d._budget_paused/_at/_reason directly
rather than declaring `global`, which correctly mutates dashboard.py's
module attributes from the route module.

dashboard.py shrinks by ~335 lines (30397 -> 30062); routes/alerts.py
is 402 lines.
